- Bugfix: sometimes ( not very often ) sortChunks would end up in an infinite loop.
There is a bug in the way chunks are sorted.  Did not find the source of
the problem but found a temporary solution which is to not to sort the
remainding chunks.  Looks like its at the most 3 chunks that don't get
sorted based on the test cases tried.  Only seems to happen on
complicated patterns.