fix issue with sorting chunks
Infinite loop was caused by don't merge paths option. This option can
cause chunks to intersect, and then seemed for the algorithm as one
inside each other, thus causing a parent-child loop.
Important to mention is the fact, that don't merge outlines is an exotic
option which is there ONLY for milling PCB's,...