optimize chunk sorting
this optimizes chunk sorting a bit, by not iterating over too many
children - some chunks could be checked many times.
For parallel operations with layers, this fixes something that was
rather a bug - because too many relations were created and this could
lead to exponential computation times relative to count of layers.