mesh modifiers option introduced problems with silhouette computation
mesh modifiers option caused zbuffer computation just because of the
mesh modifiers option. This souldn't be the wanted behaviour in my
opinion, since exact silh computation should is allready limited with
totfaces option. If you want to disable exact computation on some
meshes, expose the totfaces prop in the UI.
Also, this caused Zbuffer computation also for simplest curves cutouts,
and I consider it a bug in such cases.