Path Control  - safe Z start
Path control G64 was added.  Value above 0 will activate it.
the first motion on the machine was a fast move to 
freemovement_height = 35
G00 X0 Y0 ,Z35
this has now been changed to 
G00 Z35
X0 Y0 Z35
This will protect the bit if process stopped in the middle of milling.