numba removed, machine and operation bugfixes
Commented out `numba` import from the `__init__` for now, until the issues with crashing can be resolved.
Moved `_IS_LOADING_DEFAULTS` out of the `updateMachine` function, declared the default value = False, and grabbed it via `utils.IS_LOADING_DEFAULTS`.
(Note: It was already declared a `global` variable within the function, and I thought that would have been enough to avoid this error, but it works now)

Lastly, altered the cleanup in the `_calc_path` function in `ops.py` - previously it was checking for and deleting objects without issues. 
In Blender 4.1 deleting an object would still leave behind mesh data that retained the object name, causing the cleanup function to not recognize and update old operations.
Simply deleting the mesh data solves the issue, as it will also delete the object with it, even though the reverse is no longer true.