Preset and Preview Fix
Cutter and Machine Presets worked without issue.
Operation presets would load, but could not be saved - `AttributeError`
This was caused by 3 errors in `AddPresetCamOperation` in `init`:
- `o.movement.movement.spindle_rotation` changed to `o.movement.spindle_rotation`
- `o.optimize` and `o.optimize_threshold` changed to `o.optimisation.optimize`  and `o.optimisation.optimize_threshold`

Discovered this while trying to add a small usability patch that would add an operation if there wasn't already on in the scene, so that loading an operation preset wouldn't cause an error message.
Patch was done by adding a `setup_operation_preset` function in `utils` that would `try: except` to try to adjust an existing operation, but add one if one isn't present.
I also patched the existing operation presets, though they aren't copied to disk by default, so I added a `BoolProperty` to the `AddonPreferences` and copy logic to the `check_operations_on_load` handler so that they will only be copied once. The copy operation does not affect any additional presets the user may have, only the default set that come with the addon.

Also fixed the 'Material Preview' viewport mode. 
Previously the object would disappear when switching to that mode.
Render Engines have the option to 'borrow' Material Preview mode from EEVEE via a boolean `bl_use_eevee_viewport = True`
I set that, and added the panels to the Material tab when in Cam.